Persist bookmark each time user clicks the button and fix database models

+++ server/models/bookmark.js
@@ -1,34 +1,32 @@
 const presentationObject = require('./presentation');
 const models = require('../../database/models/index');
 
+const slides = presentationObject.getPresentation().slides;
+
 module.exports.addBookmark = function (bookmarkedSlideUrl, userId) {
-  const slides = presentationObject.getPresentation().slides;
   for (var i = 0; i < slides.length; i++) {
     if (slides[i].original === bookmarkedSlideUrl) {
       models.Bookmark.create({
-        slide_id: slideObject.dataValues.id,
-        user_id: slides[i].id
+        slide_id: slides[i].id,
+        user_id: userId
       })
       .catch(err => console.log(err));
+      break;
     }
   }
 };
 
 module.exports.removeBookmark = function (bookmarkedSlideUrl, userId) {
-  models.Slide.findOne({
-    attributes: ['id'],
-    where: {
-      image_url: bookmarkedSlideUrl
+  for (var i = 0; i < slides.length; i++) {
+    if (slides[i].original === bookmarkedSlideUrl) {
+      models.Bookmark.destroy({
+        where: {
+          slide_id: slides[i].id,
+          user_id: userId
+        }
+      })
+      .catch(err => console.log(err));
+      break;
     }
-  })
-  .then((slideObject) => {
-    models.Bookmark.destroy({
-      where: {
-        slide_id: slideObject.dataValues.id,
-        user_id: userId
-      }
-    })
-    .catch(err => console.log(err));
-  })
-  .catch(err => console.log(err));
+  }
 };
